+++ src/components/ui/modals/SafeSettingsModal.tsx
@@ -8,10 +8,21 @@ import {
   Address,
   encodeAbiParameters,
   encodeFunctionData,
+  encodePacked,
   getContract,
+  getCreate2Address,
+  keccak256,
   parseAbiParameters,
 } from 'viem';
+import {
+  linearERC20VotingWithWhitelistSetupParams,
+  linearERC721VotingWithWhitelistSetupParams,
+  linearERC721VotingWithWhitelistV1SetupParams,
+  linearERC20VotingWithWhitelistV1SetupParams,
+} from '../../../constants/params';
 import { DAO_ROUTES } from '../../../constants/routes';
+import { getRandomBytes } from '../../../helpers';
+import useFeatureFlag from '../../../helpers/environmentFeatureFlags';
 import { usePaymasterDepositInfo } from '../../../hooks/DAO/accountAbstraction/usePaymasterDepositInfo';
 import { useCurrentDAOKey } from '../../../hooks/DAO/useCurrentDAOKey';
 import { useValidationAddress } from '../../../hooks/schemas/common/useValidationAddress';
@@ -19,11 +30,13 @@ import { useNetworkEnsAddressAsync } from '../../../hooks/useNetworkEnsAddress';
 import useNetworkPublicClient from '../../../hooks/useNetworkPublicClient';
 import { useCanUserCreateProposal } from '../../../hooks/utils/useCanUserSubmitProposal';
 import { useInstallVersionedVotingStrategy } from '../../../hooks/utils/useInstallVersionedVotingStrategy';
+import { generateContractByteCodeLinear } from '../../../models/helpers/utils';
 import { SafeGeneralSettingsPage } from '../../../pages/dao/settings/general/SafeGeneralSettingsPage';
 import { useDAOStore } from '../../../providers/App/AppProvider';
 import { useNetworkConfigStore } from '../../../providers/NetworkConfig/useNetworkConfigStore';
 import { useProposalActionsStore } from '../../../store/actions/useProposalActionsStore';
 import {
+  AzoriusGovernance,
   BigIntValuePair,
   CreateProposalActionData,
   CreateProposalTransaction,
@@ -38,6 +51,7 @@ import {
   getVoteSelectorAndValidator,
 } from '../../../utils/gaslessVoting';
 import { validateENSName } from '../../../utils/url';
+import { SafePermissionsStrategyAction } from '../../SafeSettings/SafePermissionsStrategyAction';
 import { SettingsNavigation } from '../../SafeSettings/SettingsNavigation';
 import { NewSignerItem } from '../../SafeSettings/Signers/SignersContainer';
 import Divider from '../utils/Divider';
@@ -91,8 +105,15 @@ export function SafeSettingsModal({
 
   const {
     node: { safe },
-    governance: { paymasterAddress },
-    governanceContracts: { strategies, moduleAzoriusAddress },
+    governance,
+    governanceContracts: {
+      strategies,
+      moduleAzoriusAddress,
+      linearVotingErc20Address,
+      linearVotingErc721Address,
+      linearVotingErc20WithHatsWhitelistingAddress,
+      linearVotingErc721WithHatsWhitelistingAddress,
+    },
   } = useDAOStore({ daoKey });
 
   const [settingsContent, setSettingsContent] = useState(<SafeGeneralSettingsPage />);
@@ -109,7 +130,17 @@ export function SafeSettingsModal({
 
   const {
     chain: { id: chainId },
-    contracts: { keyValuePairs, accountAbstraction, paymaster, zodiacModuleProxyFactory },
+    contracts: {
+      keyValuePairs,
+      accountAbstraction,
+      paymaster,
+      zodiacModuleProxyFactory,
+      linearVotingErc20MasterCopy,
+      linearVotingErc721MasterCopy,
+      linearVotingErc20V1MasterCopy,
+      linearVotingErc721V1MasterCopy,
+      hatsProtocol,
+    },
     bundlerMinimumStake,
   } = useNetworkConfigStore();
 
@@ -174,6 +205,8 @@ export function SafeSettingsModal({
 
   const publicClient = useNetworkPublicClient();
 
+  const gaslessVotingFeatureEnabled = useFeatureFlag('flag_gasless_voting');
+
   const ethValue = {
     bigintValue: 0n,
     value: '0',
@@ -186,6 +219,7 @@ export function SafeSettingsModal({
 
     const accountAbstractionSupported = bundlerMinimumStake !== undefined;
     const stakingRequired = accountAbstractionSupported && bundlerMinimumStake > 0n;
+    const { paymasterAddress } = governance;
 
     const transactions: CreateProposalTransaction[] = [];
 
@@ -639,13 +673,263 @@ export function SafeSettingsModal({
     };
   };
 
+  const handleEditPermissions = async (updatedValues: SafeSettingsEdits) => {
+    if (!safe?.address) {
+      throw new Error('Safe address is not set');
+    }
+
+    if (!updatedValues.permissions) {
+      throw new Error('Permissions are not set');
+    }
+
+    const { proposerThreshold } = updatedValues.permissions;
+
+    if (!proposerThreshold?.bigintValue) {
+      throw new Error('Proposer threshold is not set');
+    }
+
+    let transactions: CreateProposalTransaction[] = [];
+
+    const azoriusGovernance = governance as AzoriusGovernance;
+
+    if (!moduleAzoriusAddress) {
+      throw new Error('Azorius module address is not set');
+    }
+
+    let actionType: ProposalActionType = ProposalActionType.EDIT;
+
+    if (linearVotingErc20Address) {
+      transactions = [
+        {
+          targetAddress: linearVotingErc20Address,
+          ethValue: {
+            bigintValue: 0n,
+            value: '0',
+          },
+          functionName: 'updateRequiredProposerWeight',
+          parameters: [
+            {
+              signature: 'uint256',
+              value: proposerThreshold.toString(),
+            },
+          ],
+        },
+      ];
+    } else if (linearVotingErc721Address) {
+      transactions = [
+        {
+          targetAddress: linearVotingErc721Address,
+          ethValue: {
+            bigintValue: 0n,
+            value: '0',
+          },
+          functionName: 'updateProposerThreshold',
+          parameters: [
+            {
+              signature: 'uint256',
+              value: proposerThreshold.toString(),
+            },
+          ],
+        },
+      ];
+    } else if (linearVotingErc20WithHatsWhitelistingAddress) {
+      // @todo - definitely could be more DRY here and with useCreateRoles
+      actionType = ProposalActionType.ADD;
+      const strategyNonce = getRandomBytes();
+      const linearERC20VotingMasterCopyContract = getContract({
+        abi: abis.LinearERC20Voting,
+        address: linearVotingErc20MasterCopy,
+        client: publicClient,
+      });
+
+      const { votesToken, votingStrategy } = azoriusGovernance;
+
+      if (!votesToken || !votingStrategy?.votingPeriod || !votingStrategy.quorumPercentage) {
+        throw new Error('Voting strategy or votes token not found');
+      }
+
+      const quorumDenominator = await linearERC20VotingMasterCopyContract.read.QUORUM_DENOMINATOR();
+
+      const encodedStrategyInitParams =
+        gaslessVotingFeatureEnabled && accountAbstraction
+          ? encodeAbiParameters(parseAbiParameters(linearERC20VotingWithWhitelistV1SetupParams), [
+              safe.address, // owner
+              votesToken.address, // governance token
+              moduleAzoriusAddress, // Azorius module
+              Number(votingStrategy.votingPeriod.value),
+              (votingStrategy.quorumPercentage.value * quorumDenominator) / 100n,
+              500000n,
+              hatsProtocol, // hats protocol
+              [], // whitelisted hat ids
+              accountAbstraction.lightAccountFactory, // light account factory
+            ])
+          : encodeAbiParameters(parseAbiParameters(linearERC20VotingWithWhitelistSetupParams), [
+              safe.address, // owner
+              votesToken.address, // governance token
+              moduleAzoriusAddress, // Azorius module
+              Number(votingStrategy.votingPeriod.value),
+              (votingStrategy.quorumPercentage.value * quorumDenominator) / 100n,
+              500000n,
+              hatsProtocol, // hats protocol
+              [], // whitelisted hat ids
+            ]);
+
+      const encodedStrategySetupData = encodeFunctionData({
+        abi: gaslessVotingFeatureEnabled
+          ? abis.LinearERC20VotingWithHatsProposalCreationV1
+          : abis.LinearERC20VotingWithHatsProposalCreation,
+        functionName: 'setUp',
+        args: [encodedStrategyInitParams],
+      });
+
+      const masterCopy = gaslessVotingFeatureEnabled
+        ? linearVotingErc20V1MasterCopy
+        : linearVotingErc20MasterCopy;
+
+      const strategyByteCodeLinear = generateContractByteCodeLinear(masterCopy);
+
+      const strategySalt = keccak256(
+        encodePacked(
+          ['bytes32', 'uint256'],
+          [keccak256(encodePacked(['bytes'], [encodedStrategySetupData])), strategyNonce],
+        ),
+      );
+
+      const predictedStrategyAddress = getCreate2Address({
+        from: zodiacModuleProxyFactory,
+        salt: strategySalt,
+        bytecodeHash: keccak256(encodePacked(['bytes'], [strategyByteCodeLinear])),
+      });
+
+      transactions = [
+        {
+          targetAddress: zodiacModuleProxyFactory,
+          functionName: 'deployModule',
+          ethValue: { bigintValue: 0n, value: '0' },
+          parameters: [
+            {
+              signature: 'address',
+              value: masterCopy,
+            },
+            { signature: 'bytes', value: encodedStrategySetupData },
+            { signature: 'uint256', value: strategyNonce.toString() },
+          ],
+        },
+        {
+          targetAddress: moduleAzoriusAddress,
+          functionName: 'enableStrategy',
+          ethValue: { bigintValue: 0n, value: '0' },
+          parameters: [{ signature: 'address', value: predictedStrategyAddress }],
+        },
+      ];
+    } else if (linearVotingErc721WithHatsWhitelistingAddress) {
+      actionType = ProposalActionType.ADD;
+      const strategyNonce = getRandomBytes();
+
+      const { erc721Tokens, votingStrategy } = azoriusGovernance;
+
+      if (!erc721Tokens || !votingStrategy?.votingPeriod || !votingStrategy.quorumThreshold) {
+        throw new Error('Voting strategy or NFT votes tokens not found');
+      }
+
+      const encodedStrategyInitParams =
+        gaslessVotingFeatureEnabled && accountAbstraction
+          ? encodeAbiParameters(parseAbiParameters(linearERC721VotingWithWhitelistV1SetupParams), [
+              safe.address, // owner
+              erc721Tokens.map(token => token.address), // governance token
+              erc721Tokens.map(token => token.votingWeight),
+              moduleAzoriusAddress,
+              Number(votingStrategy.votingPeriod.value),
+              proposerThreshold.bigintValue,
+              500000n,
+              hatsProtocol, // hats protocol
+              [], // whitelisted hat ids
+              accountAbstraction.lightAccountFactory, // light account factory
+            ])
+          : encodeAbiParameters(parseAbiParameters(linearERC721VotingWithWhitelistSetupParams), [
+              safe.address, // owner
+              erc721Tokens.map(token => token.address), // governance token
+              erc721Tokens.map(token => token.votingWeight),
+              moduleAzoriusAddress,
+              Number(votingStrategy.votingPeriod.value),
+              proposerThreshold.bigintValue,
+              500000n,
+              hatsProtocol, // hats protocol
+              [], // whitelisted hat ids
+            ]);
+
+      const encodedStrategySetupData = encodeFunctionData({
+        abi: gaslessVotingFeatureEnabled
+          ? abis.LinearERC20VotingWithHatsProposalCreationV1
+          : abis.LinearERC20VotingWithHatsProposalCreation,
+        functionName: 'setUp',
+        args: [encodedStrategyInitParams],
+      });
+
+      const masterCopy = gaslessVotingFeatureEnabled
+        ? linearVotingErc721V1MasterCopy
+        : linearVotingErc721MasterCopy;
+
+      const strategyByteCodeLinear = generateContractByteCodeLinear(masterCopy);
+
+      const strategySalt = keccak256(
+        encodePacked(
+          ['bytes32', 'uint256'],
+          [keccak256(encodePacked(['bytes'], [encodedStrategySetupData])), strategyNonce],
+        ),
+      );
+
+      const predictedStrategyAddress = getCreate2Address({
+        from: zodiacModuleProxyFactory,
+        salt: strategySalt,
+        bytecodeHash: keccak256(encodePacked(['bytes'], [strategyByteCodeLinear])),
+      });
+
+      transactions = [
+        {
+          targetAddress: zodiacModuleProxyFactory,
+          functionName: 'deployModule',
+          ethValue: { bigintValue: 0n, value: '0' },
+          parameters: [
+            {
+              signature: 'address',
+              value: masterCopy,
+            },
+            { signature: 'bytes', value: encodedStrategySetupData },
+            { signature: 'uint256', value: strategyNonce.toString() },
+          ],
+        },
+        {
+          targetAddress: moduleAzoriusAddress,
+          functionName: 'enableStrategy',
+          ethValue: { bigintValue: 0n, value: '0' },
+          parameters: [{ signature: 'address', value: predictedStrategyAddress }],
+        },
+      ];
+    } else {
+      throw new Error('No existing voting strategy address found');
+    }
+
+    return {
+      actionType,
+      content: (
+        <SafePermissionsStrategyAction
+          actionType={actionType}
+          proposerThreshold={proposerThreshold}
+        />
+      ),
+      transactions,
+    };
+  };
+
   const submitAllSettingsEditsProposal = async (values: SafeSettingsEdits) => {
     if (!safe?.address) {
       throw new Error('Safe address is not set');
     }
 
     resetActions();
-    const { general, multisig, azorius } = values;
+    const { general, multisig, azorius, permissions } = values;
+    console.log('submitting', { general, multisig, azorius, permissions });
     if (general) {
       const { action, title } = await handleEditGeneral(values);
 
@@ -675,6 +959,14 @@ export function SafeSettingsModal({
         content: <Text>{title}</Text>,
       });
     }
+
+    if (permissions) {
+      const action = await handleEditPermissions(values);
+
+      addAction(action);
+    }
+
+    console.log('navigating');
     navigate(DAO_ROUTES.proposalWithActionsNew.relative(addressPrefix, safe.address));
   };
